third_party/blink/web_tests/virtual/overlay-scrollbar/plugin-overlay-scrollbar-mouse-capture-expected.txt

This tests whether scrolling still works correctly when an overlay scrollbar is over a plugin. Clicking on the overlay scrollbar should cause it to activate and capture input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS eventHistory.length is 0
PASS window.scrollY > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
